Change ViewCourse.cshtml become Trainning.cshtml
