+++ HealthCommunitiesCheck2.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.11.35312.102
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HealthCommunitiesCheck2", "HealthCommunitiesCheck2.csproj", "{251BABD4-E5F1-43A0-A81E-FEE7E88CDDE3}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{251BABD4-E5F1-43A0-A81E-FEE7E88CDDE3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{251BABD4-E5F1-43A0-A81E-FEE7E88CDDE3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{251BABD4-E5F1-43A0-A81E-FEE7E88CDDE3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{251BABD4-E5F1-43A0-A81E-FEE7E88CDDE3}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {2543DCED-68F9-408D-B98D-8600CEED7F7A}
+	EndGlobalSection
+EndGlobal
