+++ Pages/ManageCourse.cshtml
@@ -41,5 +41,6 @@
 }
 else
 {
+    <a asp-page="/CreateCourse" class="btn btn-primary">Tạo khóa học</a>
     <p>No courses available!</p>
 }
