+++ Pages/Training.cshtml
@@ -3,3 +3,45 @@
 @{
 	ViewData["Title"] = "Training";
 }
+
+<h2>@ViewData["Title"]</h2>
+
+@if (Model.Courses != null && Model.Courses.Any())
+{
+    <table class="table">
+        <thead>
+            <tr>
+                <th>No.</th>
+                <th>Course Title</th>
+                <th>Description</th>
+
+                <th>Start Date</th>
+                <th>End Date</th>
+                <th>Format</th>
+                <th>Price</th>
+            </tr>
+        </thead>
+        <tbody>
+            @for (int i = 0; i < Model.Courses.Count; i++)
+            {
+                var course = Model.Courses[i];
+                <tr>
+                    <td>@(i + 1)</td>
+                    <td>@course.Title</td>
+                    <td>@course.Description</td>
+
+                    <td>@course.StartDate.ToString("MM/dd/yyyy")</td>
+                    <td>@course.EndDate.ToString("MM/dd/yyyy")</td>
+                    <td>@(course.IsOnline ? "Online" : "Offline")</td>
+                    <td>@course.Price.ToString("C")</td>
+                </tr>
+            }
+        </tbody>
+    </table>
+
+}
+else
+{
+    <p>No courses available!</p>
+}
+
