+++ Pages/ViewCourse.cshtml
@@ -1,4 +1,6 @@
 ﻿@page
-@model HealthCommunitiesCheck2.Pages.ViewCourseModel
+@model ViewCourseModel
 @{
+    ViewData["Title"] = "Course List";
 }
+
