+++ appsettings.json
@@ -7,7 +7,7 @@
     },
     "AllowedHosts": "*",
 
-    "ConnectionStrings": {
-        "DefaultConnection": "Server=MSI\\GOON;Database=HealthCommunitiesDB;User Id=sa;Password=12345;Trusted_Connection=False;MultipleActiveResultSets=true;TrustServerCertificate=True"
-    }
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=LAPTOP-QB4QC5DV;Database=HealthCommunitiesDB;User Id=sa;Password=12345;Trusted_Connection=False;MultipleActiveResultSets=true;TrustServerCertificate=True"
+  }
 }
